pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gambeson"
version = "0.1.0"
description = "Parametric CAD + FEA simulation pipeline for a modern gambeson armor system"
readme = "Modern_Gambeson_System_Whitepaper.md"
license = "CERN-OHL-S-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Gambeson Project Contributors" },
]
keywords = ["armor", "simulation", "FEA", "CAD", "parametric"]

dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
cad = [
    "cadquery>=2.4",
    "solidpython2>=2.1",
]
sim = [
    "meshio>=5.3",
    "gmsh>=4.12",
    "pyelmer>=1.0",
]
viz = [
    "pyvista>=0.43",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
all = ["gambeson[cad,sim,viz,dev]"]

[project.scripts]
gambeson = "src.pipeline.orchestrator:main"

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
